Add files via upload
51 equally sized balls are put into an urn numbered 0, 1, 2, ..., 50. You will play the
following game. You draw a number. If that number is 0 you lose and take home nothing.
Otherwise, you can take home the number of dollars on the ball OR you can put the ball back
in the hat and try again. This python script determines the optimal strategy and its expected winnings.